Add day-11 solution
I was stuck on part 2 of the problem and looked up part of the solution
on reddit (the concept on how to keep the item values (my worrying
level) low, namely the least common multiple (LCM).
By dividing by the LCM we can keep the item values low and won't alter
the test results of the monkeys.